Action queue processing refactor.
Each invocation of the middleware now checks the action queue.
If the queue is not empty, the next queued action is processed.
If the current action is interceptable, it is only processed if
the action queue is empty, otherwise it is enqueued for future
future processing. This ensures that all interceptable actions
are processed in order, even those originating from a timer.